+++ tests/sql_test_files/it_works_set_variable_null.sql
@@ -1,7 +1,0 @@
-set path = sqlpage.path();
-set url = sqlpage.set_variable('x', null);
-select 'text' as component,
-    case
-        when $url = $path THEN 'It works !'
-        else 'error: ' || $url
-    end as contents;